% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettop.R
\name{get_subclustering_features}
\alias{get_subclustering_features}
\title{get lmFit F-stat based collection of n genes most varying in mean across subclusters}
\usage{
get_subclustering_features(inlist, clname, n = 20)
}
\arguments{
\item{inlist}{list of SingleCellExperiments (SCEs) formed by scran::quickSubCluster}

\item{clname}{character(1) name of cluster SCE to assess}

\item{n}{numeric(1) number to preserve}
}
\description{
get lmFit F-stat based collection of n genes most varying in mean across subclusters
}
\examples{
data(all.sce)
scl = get_subclustering_features(all.sce, "3", 10)
names(scl)
}
